+++ src/main/java/webserver/http/request/HttpProtocol.java
@@ -1,4 +1,4 @@
-package domain;
+package webserver.http.request;
 
 import java.util.Objects;
