+++ src/main/java/webserver/http/request/QueryParameter.java
@@ -1,4 +1,4 @@
-package domain;
+package webserver.http.request;
 
 import java.util.Arrays;
 import java.util.Collections;
