+++ src/test/java/webserver/http/HttpMethodTest.java
@@ -1,4 +1,4 @@
-package domain;
+package webserver.http;
 
 import static org.assertj.core.api.Assertions.assertThat;
